+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Replace `postversion` script with `version` script for CI release.
 - Remove `package-lock.json` from root (since we use pnpm)
 - Update HTTP response check to `!==200` in `avivator/utils.js`, and changed variable name `isOffsets404` to `isOffsetsNot200`
+- Fix `onHover` function in `VivViewer` after deck.gl 8.8 upgrade.
 
 ## 0.13.1